DotEnvironmentNewLineEscaper added
