+++ src/GiGraph.Dot.Output/Options/DotSyntaxRules.Attributes.cs
@@ -1,5 +1,5 @@
-﻿using GiGraph.Dot.Output.TextEscaping.Escapers;
-using GiGraph.Dot.Output.TextEscaping.Pipelines;
+﻿using GiGraph.Dot.Output.Text.Escaping;
+using GiGraph.Dot.Output.Text.Escaping.Pipelines;
 
 namespace GiGraph.Dot.Output.Options
 {
