+++ src/GiGraph.Dot.Output/Text/Escaping/DotCurlyBracketsEscaper.cs
@@ -1,4 +1,4 @@
-﻿namespace GiGraph.Dot.Output.TextEscaping.Escapers
+﻿namespace GiGraph.Dot.Output.Text.Escaping
 {
     /// <summary>
     ///     Escapes curly brackets ({, }). Use for escaping text of record node fields.
