+++ src/GiGraph.Dot.Output/Text/Escaping/Pipelines/DotTextEscapingPipeline.Html.cs
@@ -1,6 +1,6 @@
-﻿using GiGraph.Dot.Output.TextEscaping.Escapers.Html;
+﻿using GiGraph.Dot.Output.Text.Escaping.Html;
 
-namespace GiGraph.Dot.Output.TextEscaping.Pipelines
+namespace GiGraph.Dot.Output.Text.Escaping.Pipelines
 {
     public partial class DotTextEscapingPipeline
     {
