+++ src/GiGraph.Dot.Types/Text/DotEscapeString.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using GiGraph.Dot.Output;
-using GiGraph.Dot.Output.TextEscaping.Escapers;
+using GiGraph.Dot.Output.Text.Escaping;
 
 namespace GiGraph.Dot.Types.Text
 {
