+++ src/GiGraph.Dot.Types/Text/DotUnescapedString.cs
@@ -1,5 +1,5 @@
 using System;
-using GiGraph.Dot.Output.TextEscaping.Escapers;
+using GiGraph.Dot.Output.Text.Escaping;
 
 namespace GiGraph.Dot.Types.Text
 {
